moved css folder inside src
