+++ client/src/components/account/Login.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import '/styles/Login.css';
+import '/src/styles/Login.css';
 
 function Login() {
   return (
