+++ client/src/styles/MindDeserves.css
